Make slot sizes 64 bit
Slot size was an int32_t in PublisherOptions, in the CreatePublisher
wire protocol and in ServerChannel, so the largest usable slot was
2GB - 64 bytes.  Anything larger wrapped negative and the server
rejected it with "num_slots and slot_size must be greater than 0".
INT32_MAX itself also failed, because the client 64-byte aligns the
request before sending it and the aligned value overflowed the field.

Widen slot size to int64_t through the options, the protobuf messages,
the client and server channels, the growth path and the C, Python, Rust
and Java bindings.  GetMessageBuffer() takes an int64_t max size, so the
auto-growth loop no longer truncates as it approaches 2GB.

ChannelControlBlock::max_message_size becomes a uint64_t to match; it
was silently truncating messages over 4GB.  The Rust client mirrors this
struct, so both now assert the layout of the stats block.

The Android ByteBuffer capacity is still an int, so the JNI layer
reports an error rather than returning an undefined buffer for a slot
that a JVM cannot represent.

+++ c_client/subspace.cc
@@ -557,7 +557,7 @@ SubspaceSubscriberOptions subspace_subscriber_options_default(void) {
   return options;
 }
 
-SubspacePublisherOptions subspace_publisher_options_default(int32_t slot_size,
+SubspacePublisherOptions subspace_publisher_options_default(int64_t slot_size,
                                                             int num_slots) {
   SubspacePublisherOptions options = {
       .slot_size = slot_size,
@@ -1660,7 +1660,7 @@ bool subspace_publisher_uses_split_buffers(SubspacePublisher publisher) {
          (*PublisherPtr(publisher))->UsesSplitBuffers();
 }
 
-int32_t subspace_get_publisher_slot_size(SubspacePublisher publisher) {
+int64_t subspace_get_publisher_slot_size(SubspacePublisher publisher) {
   if (publisher.publisher == nullptr) {
     return 0;
   }
@@ -1749,7 +1749,7 @@ subspace_get_publisher_virtual_memory_usage(SubspacePublisher publisher) {
 bool subspace_get_publisher_stats_counters(SubspacePublisher publisher,
                                            uint64_t *total_bytes,
                                            uint64_t *total_messages,
-                                           uint32_t *max_message_size,
+                                           uint64_t *max_message_size,
                                            uint32_t *total_drops) {
   if (total_bytes != nullptr) {
     *total_bytes = 0;
@@ -1821,7 +1821,7 @@ int subspace_get_subscriber_fd(SubspaceSubscriber subscriber) {
   return (*sub_ptr)->GetFileDescriptor().Fd();
 }
 
-int32_t subspace_get_subscriber_slot_size(SubspaceSubscriber subscriber) {
+int64_t subspace_get_subscriber_slot_size(SubspaceSubscriber subscriber) {
   if (subscriber.subscriber == nullptr) {
     return 0;
   }
