+++ shadow/shadow.h
@@ -50,7 +50,7 @@ struct ShadowSubscriber {
 struct ShadowChannel {
   std::string name;
   int channel_id = 0;
-  int slot_size = 0;
+  int64_t slot_size = 0;
   int num_slots = 0;
   uint64_t subscriber_queue_arena_size = 0;
   std::string type;
